+++ motor_abi/include/motor_abi.h
@@ -47,6 +47,8 @@ MotorController* motor_controller_new_socketcanfd(const char* channel);
 MotorController* motor_controller_new_dm_serial(const char* serial_port, uint32_t baud);
 MotorController* motor_controller_new_robstride_serial(const char* serial_port, uint32_t baud);
 MotorController* motor_controller_new_robstride_dm_serial(const char* serial_port, uint32_t baud);
+MotorController* motor_controller_new_dm_serial_mixed(const char* serial_port, uint32_t baud);
+MotorController* motor_controller_new_mixed(const char* channel);
 void motor_controller_free(MotorController* controller);
 int32_t motor_controller_poll_feedback_once(MotorController* controller);
 int32_t motor_controller_enable_all(MotorController* controller);